design feedback, downgrade warning and styles migration
