+++ packages/app-shell/src/PlanSelector/components/PlanSelectorParent.jsx
@@ -1,63 +1,35 @@
 import React, { useEffect, useState } from 'react';
-import styled from 'styled-components';
 import Text from '@bufferapp/ui/Text';
 import Switch from '@bufferapp/ui/Switch';
-import { blue } from '@bufferapp/ui/style/colors';
+import Button from '@bufferapp/ui/Button';
 import { SelectionScreen } from './SelectionScreen';
 import { Summary } from './Summary';
-import PropTypes from 'prop-types';
-
-const Container = styled.div`
-  display: flex;
-  height: 550px;
-  align-items: center;
-  border-radius: 8px;
-  box-sizing: border-box;
-  border: 1px solid grey; //REMOVE THIS BEFORE MERGING!!!!
-`;
-
-const Left = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 28px 24px 24px;
-  height: 100%;
-  justify-content: center;
-  box-sizing: border-box;
-`;
-
-const PlanSelectorHeader = styled.header`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const SwitchContainer = styled.div`
-  display: flex;
-  align-items: center;
-
-  span {
-    font-family: Roboto;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 16px;
-    letter-spacing: 0px;
-    color: ${blue};
-  }
-
-  p {
-    color: #333333;
-    font-family: Roboto, sans-serif;
-    font-size: 14px;
-    font-weight: 500;
-    line-height: 16px;
-    margin-left: 4px;
-  }
-`;
+import {
+  ButtonContainer,
+  SwitchContainer,
+  PlanSelectorHeader,
+  Right,
+  Left,
+  Container,
+} from './style';
 
 export const PlanSelectorParent = ({ planOptions }) => {
   const [selectedPlan, setselectedPlan] = useState(planOptions[0]);
   const [monthlyBilling, setBillingInterval] = useState(true);
+
+  const currentPlan = planOptions.find((option) => option.isCurrentPlan);
+  const currentPlanString = `${currentPlan.planId}_${currentPlan.planInterval}`;
+  const selectedPlanString = selectedPlan
+    ? `${selectedPlan.planId}_${selectedPlan.planInterval}`
+    : '';
+
+  const getLabel = () => {
+    return currentPlanString === selectedPlanString
+      ? 'Stay On My Current Plan'
+      : 'Confirm Plan Change';
+  };
+
+  const [label, setLabel] = useState(getLabel());
 
   const handlePlanSelection = (planString) => {
     const [selectedPlanId, selectedPlanInterval] = planString.split('_');
@@ -73,6 +45,10 @@ export const PlanSelectorParent = ({ planOptions }) => {
     const newInterval = monthlyBilling ? 'month' : 'year';
     handlePlanSelection(`${selectedPlan.planId}_${newInterval}`);
   }, [monthlyBilling]);
+
+  useEffect(() => {
+    setLabel(getLabel());
+  }, [selectedPlan]);
 
   return (
     <Container>
@@ -99,11 +75,22 @@ export const PlanSelectorParent = ({ planOptions }) => {
           monthlyBilling={monthlyBilling}
         />
       </Left>
-      <Summary
-        planOptions={planOptions}
-        selectedPlan={selectedPlan}
-        location="planSelector"
-      />
+      <Right>
+        <Summary
+          planOptions={planOptions}
+          selectedPlan={selectedPlan}
+          location="planSelector"
+        />
+        <ButtonContainer>
+          <Button
+            type="primary"
+            onClick={() => {}}
+            label={label}
+            fullWidth
+            disabled={label === 'Stay On My Current Plan'}
+          />
+        </ButtonContainer>
+      </Right>
     </Container>
   );
 };
